--- schemas.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class AskQuestion(BaseModel):
-    question: str
-    n_results: int = 3
