--- settings.py
@@ -1,12 +1,0 @@
-from pydantic_settings import BaseSettings
-
-
-class Settings(BaseSettings):
-    open_api_key: str
-
-    class Config:
-        env_file = ".env"
-        env_file_encoding = "utf-8"
-
-
-config = Settings()
