+++ src/config/__init__.py
@@ -1,0 +1,7 @@
+"""
+Configuration module for the Hubble Chatbot Backend
+"""
+
+from .settings import config, Settings
+
+__all__ = ['config', 'Settings']
