+++ src/core/__init__.py
@@ -1,1 +1,8 @@
-# Core data processing and business logic
+"""
+Core module for the Hubble Chatbot Backend
+"""
+
+from .data_processor import EndorsementDataProcessor
+from .db import get_chroma_client
+
+__all__ = ['EndorsementDataProcessor', 'get_chroma_client']
