+++ src/models/__init__.py
@@ -1,1 +1,13 @@
-# Pydantic models and data structures
+"""
+Models module for the Hubble Chatbot Backend
+"""
+
+from .api_schemas import (
+    AskQuestion, QueryRequest, QueryResponse, ProcessResponse,
+    SystemStatsResponse, CategoryResponse, ProcessesByCategoryResponse
+)
+
+__all__ = [
+    'AskQuestion', 'QueryRequest', 'QueryResponse', 'ProcessResponse',
+    'SystemStatsResponse', 'CategoryResponse', 'ProcessesByCategoryResponse'
+]
