+++ src/services/__init__.py
@@ -1,1 +1,17 @@
-# Services and external integrations
+"""
+Services module for the Hubble Chatbot Backend
+"""
+
+from .rag_system import EndorsementRAGSystem
+from .answer_service import AnswerService
+from .document_processor import DocumentProcessor
+from .embedding_service import EmbeddingService
+from .vector_store import VectorStore
+
+__all__ = [
+    'EndorsementRAGSystem',
+    'AnswerService', 
+    'DocumentProcessor',
+    'EmbeddingService',
+    'VectorStore'
+]
