--- src/utils/__init__.py
@@ -1,1 +1,0 @@
-# Utility functions and helpers
